+++ src/rastervision/filesystems/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+
+from rastervision.filesystems.filesystem import FileSystem
+from rastervision.filesystems.local_filesystem import LocalFileSystem
+from rastervision.filesystems.s3_filesystem import S3FileSystem
+from rastervision.filesystems.http_filesystem import HttpFileSystem
