+++ src/rastervision/filesystems/http_filesystem.py
@@ -1,14 +1,11 @@
-import os
 import urllib
 import urllib.request
 
-from rastervision.filesystem import FileSystem
-from typing import Union
+from rastervision.filesystems.filesystem import FileSystem
 from urllib.parse import urlparse
 
 
 class HttpFileSystem(FileSystem):
-
     def matches_uri(uri: str) -> bool:
         parsed_uri = urlparse(uri)
         return parsed_uri.scheme in ['http', 'https']
@@ -19,24 +16,12 @@ class HttpFileSystem(FileSystem):
             if response.getcode() == 200:
                 return int(response.headers['content-length']) > 0
             else:
-                return  False
+                return False
         except urllib.error.URLError:
             return False
 
-    def open(uri: str) -> Union[FileSystem, None]:
-        return None
-
-    def close(self) -> None:
+    def read(self, uri: str) -> bytearray:
         pass
 
-    def read(self) -> bytearray:
-        pass
-
-    def write(self, data: bytearray) -> int:
-        pass
-
-    def seek(self, offset: int, whence: int) -> int:
-        pass
-
-    def tell(self) -> int:
+    def write(self, uri: str, data: bytearray) -> int:
         pass
